Isolate multipart upload logic
