+++ src/toil/jobStores/aws/utils.py
@@ -22,14 +22,18 @@ import types
 
 import errno
 import uuid
+from contextlib import closing
 from ssl import SSLError
+from multiprocessing import cpu_count
 
 # Python 3 compatibility imports
 import itertools
 
 from StringIO import StringIO
 
+import boto
 from bd2k.util.exceptions import panic
+from concurrent.futures import ThreadPoolExecutor
 from six import iteritems
 
 from bd2k.util.retry import retry
@@ -250,14 +254,16 @@ def fileSizeAndTime(localFilePath):
     file_size, file_time = file_stat.st_size, file_stat.st_mtime
     return file_size, file_time
 
-def multipart(localFilePath, partSize, bucket, fileID, headers):
-    """
 
-    :param localFilePath:
-    :param partSize:
-    :param bucket:
-    :param fileID:
-    :param headers:
+def multipartUpload(localFilePath, partSize, bucket, fileID, headers):
+    """
+    Uploads a file to s3, using multipart uploading if applicable
+
+    :param str localFilePath: Path of the file to upload to s3
+    :param int partSize: max size of each part in the multipart upload, in bytes
+    :param boto.s3.Bucket bucket: the s3 bucket to upload to
+    :param str fileID: the name of the file to upload to
+    :param headers: http headers to use when uploading - generally used for encryption purposes
     :return: version of the newly uploaded file
     """
     file_size, file_time = fileSizeAndTime(localFilePath)
@@ -307,6 +313,85 @@ def multipart(localFilePath, partSize, bucket, fileID, headers):
     # Make resonably sure that the file wasn't touched during the upload
     assert fileSizeAndTime(localFilePath) == (file_size, file_time)
     return version
+
+
+def copyKeyMultipart(srcKey, dstBucketName, dstKeyName, partSize, headers=None):
+    """
+    Copies a key from a source key to a destination key in multiple parts. Note that if the
+    destination key exists it will be overwritten implicitly, and if it does not exist a new
+    key will be created. If the destination bucket does not exist an error will be raised.
+
+    :param boto.s3.key.Key srcKey: The source key to be copied from.
+    :param str dstBucketName: The name of the destination bucket for the copy.
+    :param str dstKeyName: The name of the destination key that will be created or overwritten.
+    :param int partSize: The size of each individual part, must be >= 5 MiB but large enough to
+           not exceed 10k parts for the whole file
+    :param dict headers: Any headers that should be passed.
+
+    :rtype: boto.s3.multipart.CompletedMultiPartUpload
+    :return: An object representing the completed upload.
+    """
+
+    def copyPart(partIndex):
+        if exceptions:
+            return None
+        try:
+            for attempt in retry_s3():
+                with attempt:
+                    start = partIndex * partSize
+                    end = min(start + partSize, totalSize)
+                    part = upload.copy_part_from_key(src_bucket_name=srcKey.bucket.name,
+                                                     src_key_name=srcKey.name,
+                                                     src_version_id=srcKey.version_id,
+                                                     # S3 part numbers are 1-based
+                                                     part_num=partIndex + 1,
+                                                     # S3 range intervals are closed at the end
+                                                     start=start, end=end - 1,
+                                                     headers=headers)
+        except Exception as e:
+            if len(exceptions) < 5:
+                exceptions.append(e)
+                log.error('Failed to copy part number %d:', partIndex, exc_info=True)
+            else:
+                log.warn('Also failed to copy part number %d due to %s.', partIndex, e)
+            return None
+        else:
+            log.debug('Successfully copied part %d of %d.', partIndex, totalParts)
+            # noinspection PyUnboundLocalVariable
+            return part
+
+    totalSize = srcKey.size
+    totalParts = (totalSize + partSize - 1) / partSize
+    exceptions = []
+    # We need a location-agnostic connection to S3 so we can't use the one that we
+    # normally use for interacting with the job store bucket.
+    with closing(boto.connect_s3()) as s3:
+        for attempt in retry_s3():
+            with attempt:
+                dstBucket = s3.get_bucket(dstBucketName)
+                upload = dstBucket.initiate_multipart_upload(dstKeyName, headers=headers)
+        log.info("Initiated multipart copy from 's3://%s/%s' to 's3://%s/%s'.",
+                 srcKey.bucket.name, srcKey.name, dstBucketName, dstKeyName)
+        try:
+            # We can oversubscribe cores by at least a factor of 16 since each copy task just
+            # blocks, waiting on the server. Limit # of threads to 128, since threads aren't
+            # exactly free either. Lastly, we don't need more threads than we have parts.
+            with ThreadPoolExecutor(max_workers=min(cpu_count() * 16, totalParts, 128)) as executor:
+                parts = list(executor.map(copyPart, xrange(0, totalParts)))
+                if exceptions:
+                    raise RuntimeError('Failed to copy at least %d part(s)' % len(exceptions))
+                assert len(filter(None, parts)) == totalParts
+        except:
+            with panic(log=log):
+                upload.cancel_upload()
+        else:
+            for attempt in retry_s3():
+                with attempt:
+                    completed = upload.complete_upload()
+                    log.info("Completed copy from 's3://%s/%s' to 's3://%s/%s'.",
+                             srcKey.bucket.name, srcKey.name, dstBucketName, dstKeyName)
+                    return completed
+
 
 def _put_attributes_using_post(self, domain_or_name, item_name, attributes,
                                replace=True, expected_value=None):
